ProviderGateway request parameter filtering added.
Added possibility to filter request parameters received by the
ProviderGateway. Filtering is done before sending the query to the
information system which makes it possible to modify parameter names and
values. There are two new properties for request parameter names and two
new properties for request parameter values, which makes four new
properties in total. For names and values it's possible to define filter
condition and operation in case the condition is true. New properties
are reqParamNameFilterCondition, reqParamNameFilterOperation,
reqParamValueFilterCondition and reqParamValueFilterOperation. #23

New services added to ProviderGateway.