Added support for payload enryption.
REST Gateway can now be configured to encrypt/decrypt the message
payload. Payload is encrypted/decrypted using one time symmetric AES key
that is generated automatically for each message to be encrypted. The
symmetric one time key is then encrypted using the recipients public key
and it's sent together with the encrypted data and initialization vector
(IV). The recipient decrypts the one time symmetric key with the private
key and then the message payload with the symmetric key. However, REST
Gateway does not implement the distribution of private/public keys -
that must be handled through other channels.